requirements-python-ansible-lint.txt
#####################
# Main dependencies #
#####################

ansible-lint==25.8.2


#####################################################################
# Sub-dependencies, frozen here to ensure consistency across setups #
#####################################################################

ansible-compat==25.8.1
attrs==25.3.0
black==25.1.0
bracex==2.6
click==8.2.1
filelock==3.19.1
importlib_metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
pathspec==0.12.1
platformdirs==4.4.0
referencing==0.36.2
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
subprocess-tee==0.4.2
typing_extensions==4.15.0
wcmatch==10.1
yamllint==1.37.1
zipp==3.23.0
